refactor/styling: starts & menu
-Refactor stars function into a component & apply it it to both lessons page and games page. -Add styling to stars so that it displaysmore visibly over buttons. -Add styling to buttons so that they are highlighted in blue when hovered. -Add icons to side menu on lessons page and style it accordingly. -Move lessons data into it's own component so that it can be used by both lessons and lesson page.
