+++ client/public/sitemap.xml
@@ -11,12 +11,22 @@
   </url>
 
   <url>
+    <loc>https://freetypingcamp.com/lessons/lesson/1/sec-1/lvl-1</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
     <loc>https://freetypingcamp.com/games</loc>
     <lastmod>2023-12-08T22:08:00+00:00</lastmod>
   </url>
 
   <url>
-    <loc>https://freetypingcamp.com/articles</loc>
+    <loc>https://freetypingcamp.com/games/calculator</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/learn</loc>
     <lastmod>2023-12-08T22:08:00+00:00</lastmod>
   </url>
 
@@ -29,12 +39,7 @@
     <loc>https://freetypingcamp.com/register</loc>
     <lastmod>2023-12-08T22:08:00+00:00</lastmod>
   </url>
-  
-  <url>
-    <loc>https://freetypingcamp.com/speed-calculator</loc>
-    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
-  </url>
-  
+    
   <url>
     <loc>https://freetypingcamp.com/privacypolicy</loc>
     <lastmod>2023-12-08T22:08:00+00:00</lastmod>
