+++ client/src/assets/sitemap.xml
@@ -1,8 +1,58 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
-  <loc>https://freetypingcamp.com/</loc>
-  <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+    <loc>https://freetypingcamp.com/</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+  
+  <url>
+    <loc>https://freetypingcamp.com/lessons</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/lessons/lesson/1/sec-1/lvl-1</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/games</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/games/calculator</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/learn</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/login</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/register</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+    
+  <url>
+    <loc>https://freetypingcamp.com/privacypolicy</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/cookiespolicy</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
+  </url>
+
+  <url>
+    <loc>https://freetypingcamp.com/termsofservice</loc>
+    <lastmod>2023-12-08T22:08:00+00:00</lastmod>
   </url>
 </urlset>
