+++ client/src/data/LessonsData.ts
@@ -1,0 +1,684 @@
+function LessonsData() {
+  //Array of obects used to manage lessons page structure
+  const pageData = [
+    {
+      lessonId: "beginner-id",
+      lessonIcon: "face",
+      lessonTitle: "Beginner",
+      lessonData: [
+        {
+          sectionTitle: "Home Row Left Hand",
+          sectionId: "home-row-left-id",
+          sectionData: [
+            {
+              id: "as",
+              levelTitle: "as",
+            },
+            {
+              id: "ad",
+              levelTitle: "ad",
+            },
+            {
+              id: "af",
+              levelTitle: "af",
+            },
+            {
+              id: "sd",
+              levelTitle: "sd",
+            },
+            {
+              id: "fd",
+              levelTitle: "fd",
+            },
+            {
+              id: "asd",
+              levelTitle: "asd",
+            },
+            {
+              id: "fds",
+              levelTitle: "fds",
+            },
+            {
+              id: "asdf",
+              levelTitle: "asdf",
+            },
+            {
+              id: "asdf-capital",
+              levelTitle: "ASDF",
+            },
+            {
+              id: "asdfasdf",
+              levelTitle: "ASDFasdf",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Home Row Right Hand",
+          sectionId: "home-row-right-id",
+          sectionData: [
+            {
+              id: "jk",
+              levelTitle: "jk",
+            },
+            {
+              id: "jl",
+              levelTitle: "jl",
+            },
+            {
+              id: "j;",
+              levelTitle: "j;",
+            },
+            {
+              id: "kl",
+              levelTitle: "kl",
+            },
+            {
+              id: ";k",
+              levelTitle: ";k",
+            },
+            {
+              id: "jkl",
+              levelTitle: "jkl",
+            },
+            {
+              id: ";lk",
+              levelTitle: ";lk",
+            },
+            {
+              id: "jkl:",
+              levelTitle: "JKL:",
+            },
+            {
+              id: "jkl:jkl;",
+              levelTitle: "JKL:jkl;",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Home Row",
+          sectionId: "home-row-id",
+          sectionData: [
+            {
+              id: "asjk",
+              levelTitle: "asjk",
+            },
+            {
+              id: "adjl",
+              levelTitle: "adjl",
+            },
+            {
+              id: "afj;",
+              levelTitle: "afj;",
+            },
+            {
+              id: "sdkl",
+              levelTitle: "sdkl",
+            },
+            {
+              id: "fs;k",
+              levelTitle: "fs;k",
+            },
+            {
+              id: "asdjkl",
+              levelTitle: "asdjkl",
+            },
+            {
+              id: "fds;lk",
+              levelTitle: "fds;lk",
+            },
+            {
+              id: "asdfjkl:",
+              levelTitle: "ASDFJKL:",
+            },
+            {
+              id: "asdfasdfjkl:jkl;",
+              levelTitle: "ASDFasdfJKL:jkl;",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Top Row Left Hand",
+          sectionId: "top-row-left-id",
+          sectionData: [
+            {
+              id: "qw",
+              levelTitle: "qw",
+            },
+            {
+              id: "qe",
+              levelTitle: "qe",
+            },
+            {
+              id: "qr",
+              levelTitle: "qr",
+            },
+            {
+              id: "qt",
+              levelTitle: "qt",
+            },
+            {
+              id: "wer",
+              levelTitle: "wer",
+            },
+            {
+              id: "qwe",
+              levelTitle: "qwe",
+            },
+            {
+              id: "ert",
+              levelTitle: "ert",
+            },
+            {
+              id: "qwert",
+              levelTitle: "qwert",
+            },
+            {
+              id: "qwert-capital",
+              levelTitle: "QWERT",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Top Row Right Hand",
+          sectionId: "top-row-right-id",
+          sectionData: [
+            {
+              id: "yu",
+              levelTitle: "yu",
+            },
+            {
+              id: "yi",
+              levelTitle: "yi",
+            },
+            {
+              id: "yo",
+              levelTitle: "yo",
+            },
+            {
+              id: "yp",
+              levelTitle: "yp",
+            },
+            {
+              id: "uio",
+              levelTitle: "uio",
+            },
+            {
+              id: "yui",
+              levelTitle: "yui",
+            },
+            {
+              id: "iop",
+              levelTitle: "iop",
+            },
+            {
+              id: "yuiop",
+              levelTitle: "yuiop",
+            },
+            {
+              id: "yuiop-capital",
+              levelTitle: "YUIOP",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Top Row",
+          sectionId: "top-row-id",
+          sectionData: [
+            {
+              id: "qwyu",
+              levelTitle: "qwyu",
+            },
+            {
+              id: "qeyi",
+              levelTitle: "qeyi",
+            },
+            {
+              id: "qryo",
+              levelTitle: "qryo",
+            },
+            {
+              id: "qtyp",
+              levelTitle: "qtyp",
+            },
+            {
+              id: "weruio",
+              levelTitle: "weruio",
+            },
+            {
+              id: "qweyui",
+              levelTitle: "qweyui",
+            },
+            {
+              id: "ertiop",
+              levelTitle: "ertiop",
+            },
+            {
+              id: "qwertyuiop",
+              levelTitle: "qwertyuiop",
+            },
+            {
+              id: "qwertyuiop-capital",
+              levelTitle: "QWERTYUIOP",
+            },
+            {
+              id: "qwertqwertyuiopyuiop",
+              levelTitle: "QWERTqwertYUIOPyuiop",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "intermediate-id",
+      lessonIcon: "face",
+      lessonTitle: "Intermediate",
+      lessonData: [
+        {
+          sectionTitle: "Bottom Row Left Hand",
+          sectionId: "bottom-row-left-id",
+          sectionData: [
+            {
+              id: "zx",
+              levelTitle: "zx",
+            },
+            {
+              id: "zc",
+              levelTitle: "zc",
+            },
+            {
+              id: "zv",
+              levelTitle: "zv",
+            },
+            {
+              id: "zb",
+              levelTitle: "zb",
+            },
+            {
+              id: "xcv",
+              levelTitle: "xcv",
+            },
+            {
+              id: "zxc",
+              levelTitle: "zxc",
+            },
+            {
+              id: "cvb",
+              levelTitle: "cvb",
+            },
+            {
+              id: "zxcvb",
+              levelTitle: "zxcvb",
+            },
+            {
+              id: "zxcvb-capital",
+              levelTitle: "ZXCVB",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Bottom Row Right Hand",
+          sectionId: "bottom-row-right-id",
+          sectionData: [
+            {
+              id: "nm",
+              levelTitle: "nm",
+            },
+            {
+              id: "n,",
+              levelTitle: "n,",
+            },
+            {
+              id: "n.",
+              levelTitle: "n.",
+            },
+            {
+              id: "n/",
+              levelTitle: "n/",
+            },
+            {
+              id: "m,.",
+              levelTitle: "m,.",
+            },
+            {
+              id: "nm,",
+              levelTitle: "nm,",
+            },
+            {
+              id: ",./",
+              levelTitle: ",./",
+            },
+            {
+              id: "nm,./",
+              levelTitle: "nm,./",
+            },
+            {
+              id: "nm<>?",
+              levelTitle: "NM<>?",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Bottom Row",
+          sectionId: "bottom-row-id",
+          sectionData: [
+            {
+              id: "zxnm",
+              levelTitle: "zxnm",
+            },
+            {
+              id: "zcn,",
+              levelTitle: "zcn,",
+            },
+            {
+              id: "zvn.",
+              levelTitle: "zvn.",
+            },
+            {
+              id: "zbn/",
+              levelTitle: "zbn/",
+            },
+            {
+              id: "xcvm,.",
+              levelTitle: "xcvm,.",
+            },
+            {
+              id: "zxcnm,",
+              levelTitle: "zxcnm,",
+            },
+            {
+              id: "cvb,./",
+              levelTitle: "cvb,./",
+            },
+            {
+              id: "zxcvbnm,./",
+              levelTitle: "zxcvbnm,./",
+            },
+            {
+              id: "zxcvbnm,./-capital",
+              levelTitle: "ZXCVBNM,./",
+            },
+            {
+              id: "zxcvbzxcvbnm,./nm,./",
+              levelTitle: "ZXCVBzxcvbNM,./nm,./",
+            },
+          ],
+        },
+        {
+          sectionTitle: "English Words",
+          sectionId: "all-three-rows-id",
+          sectionData: [
+            {
+              id: "lower-case-1",
+              levelTitle: "lower case: 🫲",
+            },
+            {
+              id: "lower-case-2",
+              levelTitle: "lower case: 🫱",
+            },
+            {
+              id: "lower-case-3",
+              levelTitle: "lower case: 🙌",
+            },
+            {
+              id: "upper-case-1",
+              levelTitle: "UPPER CASE: 🫲",
+            },
+            {
+              id: "upper-case-2",
+              levelTitle: "UPPER CASE: 🫱",
+            },
+            {
+              id: "upper-case-3",
+              levelTitle: "UPPER CASE: 🙌",
+            },
+            {
+              id: "camel-case-1",
+              levelTitle: "CaMeL CaSe: 🫲",
+            },
+            {
+              id: "camel-case-2",
+              levelTitle: "CaMeL CaSe: 🫱",
+            },
+            {
+              id: "camel-case-3",
+              levelTitle: "CaMeL CaSe: 🙌",
+            },
+            {
+              id: "pascal-case-1",
+              levelTitle: "Pascal Case: 🫲",
+            },
+            {
+              id: "pascal-case-2",
+              levelTitle: "Pascal Case: 🫱",
+            },
+            {
+              id: "pascal-case-3",
+              levelTitle: "Pascal Case: 🙌",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "advanced-id",
+      lessonIcon: "face",
+      lessonTitle: "Advanced",
+      lessonData: [
+        {
+          sectionTitle: "Number Row",
+          sectionId: "number-row-id",
+          sectionData: [
+            {
+              id: "number-row-left",
+              levelTitle: "123456",
+            },
+            {
+              id: "number-row-right",
+              levelTitle: "7890-=",
+            },
+            {
+              id: "number-row-full",
+              levelTitle: "1234567890-=",
+            },
+            {
+              id: "number-row-left-1",
+              levelTitle: "!@#$%^",
+            },
+            {
+              id: "number-row-right-1",
+              levelTitle: "&*()_+",
+            },
+            {
+              id: "number-left-full-1",
+              levelTitle: "123456!@#$%^",
+            },
+            {
+              id: "number-right-full-1",
+              levelTitle: "7890-=&*()_+",
+            },
+            {
+              id: "number-full-1",
+              levelTitle: "Full Number Row",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Brackets",
+          sectionId: "brackets-id",
+          sectionData: [
+            {
+              id: "brackets-1",
+              levelTitle: "asdfjkl;{}",
+            },
+            {
+              id: "brackets-2",
+              levelTitle: "asdfjkl;[]",
+            },
+            {
+              id: "brackets-3",
+              levelTitle: "asdfjkl;{}[]",
+            },
+            {
+              id: "brackets-4",
+              levelTitle: "asdfjkl;()",
+            },
+            {
+              id: "brackets-5",
+              levelTitle: "asdfjkl;{}[]()",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Symbols",
+          sectionId: "symbols-id",
+          sectionData: [
+            {
+              id: "mixed-case-1",
+              levelTitle: "MiXed CasE: 🫲",
+            },
+            {
+              id: "mixed-case-2",
+              levelTitle: "MiXed CasE: 🫱",
+            },
+            {
+              id: "mixed-case-3",
+              levelTitle: "MiXed CasE: 🙌",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Letters, Numbers, & Symbols",
+          sectionId: "letters-nums-symbols-id",
+          sectionData: [
+            {
+              id: "mixed-case-1",
+              levelTitle: "MiXed CasE: 🫲",
+            },
+            {
+              id: "mixed-case-2",
+              levelTitle: "MiXed CasE: 🫱",
+            },
+            {
+              id: "mixed-case-3",
+              levelTitle: "MiXed CasE: 🙌",
+            },
+          ],
+        },
+        {
+          sectionTitle: "Tricky Words",
+          sectionId: "tricky-words-id",
+          sectionData: [
+            {
+              id: "tricky-words",
+              levelTitle: "Words",
+            },
+            {
+              id: "tricky-mixed-words",
+              levelTitle: "MiXed CaSe",
+            },
+            {
+              id: "tricky-words-symbols",
+              levelTitle: "Words & Symbols",
+            },
+            {
+              id: "tricky-mixed-words-symbols",
+              levelTitle: "MiXed CaSe & Symbols",
+            },
+            {
+              id: "mixed-case-3",
+              levelTitle: "MiXed WorDs & All",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "graduation-id",
+      lessonIcon: "graduationHat",
+      lessonTitle: "You Made It",
+      lessonData: [
+        {
+          sectionTitle: "Graduation",
+          sectionId: "home-row-left-id",
+          sectionData: [
+            {
+              id: "congratulations",
+              levelTitle: "Congratulations!",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "quotes-id",
+      lessonIcon: "quote",
+      lessonTitle: "Quotes",
+      lessonData: [
+        {
+          sectionTitle: "Quote Section",
+          sectionId: "quote-lesson-id",
+          sectionData: [
+            {
+              id: "quote-1",
+              levelTitle: "Quote 1",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "animals-id",
+      lessonIcon: "paw",
+      lessonTitle: "Animals",
+      lessonData: [
+        {
+          sectionTitle: "Animal 1",
+          sectionId: "Animal-1-id",
+          sectionData: [
+            {
+              id: "first-animal",
+              levelTitle: "🦑 First Animal",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "biology-id",
+      lessonIcon: "microscope",
+      lessonTitle: "Biology",
+      lessonData: [
+        {
+          sectionTitle: "First Biology Section",
+          sectionId: "first-bio-lesson-id",
+          sectionData: [
+            {
+              id: "bio-text-1",
+              levelTitle: "Bio Text 1",
+            },
+          ],
+        },
+      ],
+    },
+    {
+      lessonId: "novels-id",
+      lessonIcon: "book",
+      lessonTitle: "Novels",
+      lessonData: [
+        {
+          sectionTitle: "First Book Theme",
+          sectionId: "first-book-theme-id",
+          sectionData: [
+            {
+              id: "first-book",
+              levelTitle: "First Book",
+            },
+          ],
+        },
+      ],
+    },
+  ];
+  return pageData;
+}
+
+export default LessonsData;
