+++ client/src/index.css
@@ -1,4 +1,3 @@
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -29,7 +28,6 @@
     url(./assets/fonts/Lora-Bold.woff2) format("woff2");
   font-display: swap;
 }
-
 
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
